- renamed updateUser -> updateLaunchDarkly - used getter instead of computing KC roles again - combined updateUser + updateLaunchDarkly - renamed featureFlags.ts -> launchDarkly.ts - misc cleanup
