+++ app/src/plugins/index.ts
@@ -1,7 +1,7 @@
 export * from './interceptors' // must be first
 export * from './authHelper'
-export * from './featureFlags'
 export * from './getConfig'
+export * from './launchDarkly'
 export * from './navigate'
 export * from './sleep'
 export * from './utilities'
