+++ app/src/views/auth/Signin.vue
@@ -36,7 +36,7 @@ export default class Signin extends Mixins(LoadKeycloakRolesMixin, NrAffiliation
 
     // now that the user is logged in, load Keycloak roles and update LaunchDarkly
     this.loadKeycloakRoles()
-    await this.updateUser()
+    await this.updateLaunchDarkly()
 
     // go to main app page
     await this.$router.push('/')
